% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{iconic_taxa}
\alias{iconic_taxa}
\title{Plot number of observations and species per iconic taxonomic group}
\usage{
iconic_taxa(df)
}
\arguments{
\item{df}{An \code{sf} tibble of iNaturalist occurrences containing at least
the columns \code{taxon.iconic_taxon_name} and \code{name}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object printed to the active device.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a tibble of iNaturalist occurrences and produces a grouped bar chart
showing the total number of observations and unique species for each iconic
taxonomic group as defined by iNaturalist.
}
\examples{
\dontrun{
#' # Download taxa occurrences from iNaturalist using ReLTER's
# get_site_speciesOccurrences() function
# e.g. Montagna di Torricchio eLTER site
deimsid <- "https://deims.org/6b62feb2-61bf-47e1-b97f-0e909c408db8"
site_boundary <- ReLTER::get_site_info(deimsid = deimsid)

iNat_occ_eLTER_site <- ReLTER::get_site_speciesOccurrences(
  deimsid = deimsid,
  list_DS = "inat",
  show_map = TRUE,
  limit = 50
)
occ_eLTER_enrich <- add_iucn_to_occ(
  occ_eLTER = iNat_occ_eLTER_site$inat
) |>
  add_nativeness_to_occ(
    country = site_boundary$country
  ) |>
  add_eunis_legal_to_occ()
  
iconic_taxa(occ_eLTER_enrich)
}

}
\seealso{
\code{\link{top_n_species}}, \code{\link{obs_per_year}}
}
\author{
Alessandro Oggioni, PhD \email{alessandro.oggioni@cnr.it}

Alice Lenzi, phD \email{alice.lenzi@crea.gov.it}
}
