% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ_eLTER_legal}
\alias{occ_eLTER_legal}
\title{Example iNaturalist occurrences enriched with conservation information}
\format{
An \code{sf} tibble with the following key columns:
\describe{
\item{name}{Scientific name}
\item{taxon.iconic_taxon_name}{iNaturalist iconic taxon group}
\item{status_IUCN}{list-column of IUCN conservation status tibbles}
\item{establishmentMeans}{list-column of establishment means tibbles}
\item{directive}{EU directive name}
\item{annex}{EU directive annex}
}
}
\source{
iNaturalist via \code{ReLTER::get_site_speciesOccurrences()}
}
\usage{
occ_eLTER_legal
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A dataset of iNaturalist occurrences from the Montagna di Torricchio
eLTER site, enriched with IUCN conservation status, establishment means,
and EUNIS legal information.
}
\keyword{datasets}
