% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_boundary}
\alias{site_boundary}
\title{eLTER site boundary for Montagna di Torricchio}
\format{
An \code{sf} object with site metadata and boundary polygon.
}
\source{
\url{https://deims.org/6b62feb2-61bf-47e1-b97f-0e909c408db8}
}
\usage{
site_boundary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

An \code{sf} object representing the boundary of the Montagna di
Torricchio eLTER site retrieved via \code{ReLTER::get_site_info()}.
}
\keyword{datasets}
