% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{top_n_species}
\alias{top_n_species}
\title{Plot top N most observed species}
\usage{
top_n_species(df)
}
\arguments{
\item{df}{An \code{sf} tibble of iNaturalist occurrences containing at least
the columns \code{name} and \code{taxon.iconic_taxon_name}.}

\item{n}{Integer. Number of top species to display. Default is \code{10}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object printed to the active device.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a tibble of iNaturalist occurrences and produces a horizontal bar
chart of the top \code{n} most observed species, coloured by iconic taxon
group. If fewer than \code{n} species have more than one observation, all
species are shown.
}
\examples{
\dontrun{
top_n_species(occ_eLTER_legal)
top_n_species(occ_eLTER_legal, n = 20)
}

}
\seealso{
\code{\link{iconic_taxa}}, \code{\link{obs_per_year}}
}
\author{
Alessandro Oggioni, PhD \email{alessandro.oggioni@cnr.it}

Alice Lenzi, phD \email{alice.lenzi@crea.gov.it}
}
