% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_iucn_to_obs}
\alias{add_iucn_to_obs}
\title{Enrich iNaturalist specific project observations with IUCN conservation status}
\usage{
add_iucn_to_obs(project_name)
}
\arguments{
\item{project_name}{A \code{string} value representing the name of the
iNaturalist project.}
}
\value{
A \code{data.frame} object representing the iNaturalist project
observations, as well as the conservation status of the taxa.
The structure of the \code{data.frame} follows the original structure as obtained
by the \code{rinat::get_inat_obs_project()} function where the type arguments is
set to "observations".
The content of the \code{data.frame} returned is enriched with the following
columns:
\itemize{
\item \code{info_title_proj}: A \code{character} value representing the title of the
project, as obtained by the \code{rinat::get_inat_obs_project()} function where
the type arguments is set to "info";
\item \code{info_slug_proj}: A \code{character} value representing the slug of the
project;
\item \code{info_taxa_num_proj}: A \code{numeric} value representing the number of taxa
in the project;
\item \code{info_place_uuid_proj}: A \code{character} value representing the place uuid
of the project within iNaturalist system;
\item \code{status}: A \code{character} value representing the conservation status of the
taxa by parsing the Taxa iNaturalist API JSON response, based on the
\code{taxon.id} value and only if the authority is the 'IUCN Red List'. In the
case the JSON parsing fails, it will returns an Error message and fill this
column with NA.
The function also prints some console messages to inform the user about the
status results for each \code{taxon.id}, as well as the number of taxa that
encountered a JSON parsing error.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function enriches all the iNaturalist project observations with the
IUCN Red List conservation status thanks to the \code{get_conservation_status()}
function.
The observations are filtered to include only 'Research grade' data also
that meet the following a valid date, an 'open' geographic location, the
presence of a photo or sound, and the exclusion of captive or cultivated
organisms.
}
\examples{
\dontrun{
## Not run:
add_iucn_to_obs(project_name = "LTER site Montagna di Torricchio")
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD (2023) \email{alessandro.oggioni@cnr.it}
}
