% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_eunis_legal_info}
\alias{get_eunis_legal_info}
\title{Get EUNIS Legal Information for a Species Using iNaturalist Taxon ID}
\usage{
get_eunis_legal_info(taxon.id)
}
\arguments{
\item{taxon.id}{Integer. The taxon ID of the species in iNaturalist.}
}
\value{
A tibble with the following columns:
\describe{
\item{taxon.id}{iNaturalist taxon ID provided as input}
\item{scientific_name}{Scientific name of the species retrieved from iNaturalist}
\item{\verb{Legal text}}{Legal directive text from EUNIS (92/43/EEC or 2009/147/EC)}
\item{Annex}{Annex information from EUNIS table}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function takes a \code{taxon.id} from iNaturalist, retrieves the corresponding
scientific name from the iNaturalist API, searches the EUNIS database, and extracts
the legal information related to the EU Habitats Directive (92/43/EEC) and Birds Directive (2009/147/EC),
as delivered by data retrieved from the \href{https://eunis.eea.europa.eu/}{EUNIS} species database.
}
\examples{
\dontrun{
## Not run:
get_eunis_legal_info(taxon.id = 61749)
}
## End (Not run)

}
\author{
Alessandro Oggioni, PhD (2023) \email{alessandro.oggioni@cnr.it}
}
