% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReLTER.inatEnrich-package.R
\docType{package}
\name{ReLTER.inatEnrich-package}
\alias{ReLTER.inatEnrich}
\alias{ReLTER.inatEnrich-package}
\title{ReLTER.inatEnrich: ReLTER.inatEnrich: Enriching iNaturalist occurrences for eLTER-RI sites}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oggioniale/ReLTER.inatEnrich}
  \item \url{https://oggioniale.github.io/ReLTER.inatEnrich/}
  \item Report bugs at \url{https://github.com/oggioniale/ReLTER.inatEnrich/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Oggioni \email{alessandro.oggioni@cnr.it} (\href{https://orcid.org/0000-0002-7997-219X}{ORCID}) [funder]

Authors:
\itemize{
  \item Alice Lenzi \email{alice.lenzi@crea.gov.it} (\href{https://orcid.org/0000-0002-3456-5726}{ORCID}) [contributor]
  \item Alessandro Campanaro \email{alessandro.campanaro@crea.gov.it} (\href{https://orcid.org/0000-0003-4043-6362}{ORCID}) [contributor]
  \item Caterina Bergami \email{caterina.bergami@cnr.it} (\href{https://orcid.org/0000-0002-5284-1317}{ORCID}) [contributor]
}

}
\keyword{internal}
