% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReLTER.inatEnrich-package.R
\docType{package}
\name{ReLTER.inatEnrich-package}
\alias{ReLTER.inatEnrich}
\alias{ReLTER.inatEnrich-package}
\title{ReLTER.inatEnrich: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oggioniale/ReLTER.inatEnrich}
  \item \url{https://oggioniale.github.io/ReLTER.inatEnrich/}
  \item Report bugs at \url{https://github.com/oggioniale/ReLTER.inatEnrich/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Oggioni \email{alessandro.oggioni@cnr.it} (\href{https://orcid.org/0000-0002-7997-219X}{ORCID}) [funder]

}
\keyword{internal}
